:recycle: api: removing 'for-admin' in filenames
