+++ api/lib/application/complementary-certifications/complementary-certification-controller.js
@@ -1,6 +1,6 @@
 import { usecases } from '../../domain/usecases/index.js';
 import * as complementaryCertificationSerializer from '../../infrastructure/serializers/jsonapi/complementary-certification-serializer.js';
-import * as targetProfilesAttachableSerializer from '../../infrastructure/serializers/jsonapi/target-profiles-attachable-for-admin-serializer.js';
+import * as attachableTargetProfilesSerializer from '../../infrastructure/serializers/jsonapi/attachable-target-profiles-serializer.js';
 
 const findComplementaryCertifications = async function () {
   const complementaryCertifications = await usecases.findComplementaryCertifications();
@@ -17,8 +17,8 @@ const getComplementaryCertificationTargetProfileHistory = async function (reques
 
 const searchAttachableTargetProfilesForComplementaryCertifications = async function (request) {
   const searchTerm = request.query.searchTerm;
-  const targetProfilesAttachable = await usecases.searchAttachableTargetProfilesForAdmin({ searchTerm });
-  return targetProfilesAttachableSerializer.serialize(targetProfilesAttachable);
+  const attachableTargetProfiles = await usecases.searchAttachableTargetProfiles({ searchTerm });
+  return attachableTargetProfilesSerializer.serialize(attachableTargetProfiles);
 };
 
 const complementaryCertificationController = {
