--- api/lib/domain/usecases/search-attachable-target-profiles-for-admin.js
@@ -1,8 +1,0 @@
-const searchAttachableTargetProfilesForAdmin = async function ({
-  searchTerm,
-  targetProfileAttachableForAdminRepository,
-}) {
-  return targetProfileAttachableForAdminRepository.find({ searchTerm });
-};
-
-export { searchAttachableTargetProfilesForAdmin };
