+++ api/lib/domain/usecases/search-attachable-target-profiles.js
@@ -1,0 +1,5 @@
+const searchAttachableTargetProfiles = async function ({ searchTerm, attachableTargetProfileRepository }) {
+  return attachableTargetProfileRepository.find({ searchTerm });
+};
+
+export { searchAttachableTargetProfiles };
