+++ api/tests/unit/infrastructure/serializers/jsonapi/attachable-target-profiles-serializer_test.js
@@ -1,0 +1,31 @@
+import { expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/attachable-target-profiles-serializer.js';
+import { AttachableTargetProfile } from '../../../../../lib/domain/models/AttachableTargetProfile.js';
+
+describe('Unit | Serializer | JSONAPI | attachable-target-profiles-serializer', function () {
+  describe('#serialize', function () {
+    it('should serialize attachable target profiles to JSONAPI', function () {
+      // given
+      const attachableTargetProfile = new AttachableTargetProfile({
+        id: 1,
+        name: 'AttachableTargetProfiles',
+      });
+
+      // when
+      const serializedAttachableTargetProfiles = serializer.serialize([attachableTargetProfile]);
+
+      // then
+      expect(serializedAttachableTargetProfiles).to.deep.equal({
+        data: [
+          {
+            attributes: {
+              name: 'AttachableTargetProfiles',
+            },
+            id: '1',
+            type: 'attachable-target-profiles',
+          },
+        ],
+      });
+    });
+  });
+});
